Move the remaining handlers into the handler package
