+++ src/main/java/com/yelp/nrtsearch/server/grpc/LuceneServer.java
@@ -27,13 +27,13 @@ import com.google.protobuf.Empty;
 import com.google.protobuf.util.JsonFormat;
 import com.yelp.nrtsearch.server.config.LuceneServerConfiguration;
 import com.yelp.nrtsearch.server.config.QueryCacheConfig;
-import com.yelp.nrtsearch.server.luceneserver.AddReplicaHandler;
-import com.yelp.nrtsearch.server.luceneserver.CopyFilesHandler;
-import com.yelp.nrtsearch.server.luceneserver.GetNodesInfoHandler;
-import com.yelp.nrtsearch.server.luceneserver.NewNRTPointHandler;
-import com.yelp.nrtsearch.server.luceneserver.RecvCopyStateHandler;
-import com.yelp.nrtsearch.server.luceneserver.ReplicaCurrentSearchingVersionHandler;
-import com.yelp.nrtsearch.server.luceneserver.WriteNRTPointHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.AddReplicaHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.CopyFilesHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.GetNodesInfoHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.NewNRTPointHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.RecvCopyStateHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.ReplicaCurrentSearchingVersionHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.WriteNRTPointHandler;
 import com.yelp.nrtsearch.server.luceneserver.analysis.AnalyzerCreator;
 import com.yelp.nrtsearch.server.luceneserver.concurrent.ThreadPoolExecutorFactory;
 import com.yelp.nrtsearch.server.luceneserver.custom.request.CustomRequestProcessor;
